Add pbug::java_agent defined type
This commit adds a defined type for managing the presence or absence of Java
agents attached to a service. The implementation uses an ini_subsetting
resource to add or remove a value from and environment variable that defines
the CLI arguments passed to the service. The type can handle agents distributed
as JAR files or shared object libraries.